+++ LlamaCppLib/LlmEngine.cs
@@ -134,18 +134,18 @@ namespace LlamaCppLib
         }
 
         [UnmanagedCallersOnly(CallConvs = [typeof(CallConvCdecl)])]
-        private static unsafe byte _ProgressCallback(float progress, void* state)
+        private static unsafe sbyte _ProgressCallback(float progress, void* state)
         {
             var callback = (Action<float>?)GCHandle.FromIntPtr(new(state)).Target;
             callback?.Invoke(progress * 100);
-            return (byte)(true ? 1 : 0);
+            return true ? 1 : 0;
         }
 
         [UnmanagedCallersOnly(CallConvs = [typeof(CallConvCdecl)])]
-        static unsafe byte AbortCallback(void* state)
+        static unsafe sbyte AbortCallback(void* state)
         {
             var cancellationToken = (CancellationToken?)GCHandle.FromIntPtr(new(state)).Target;
-            return (byte)(cancellationToken?.IsCancellationRequested ?? false ? 1 : 0);
+            return (sbyte)(cancellationToken?.IsCancellationRequested ?? false ? 1 : 0);
         }
 
         private void _StartAsync()
